+++ dpat/mil/models/ccmil.py
@@ -51,7 +51,7 @@ class CCMIL(VarAttention):
         return features
 
     @cache
-    def _calculate_llm_hidden_dim_size(self):
+    def _calculate_llm_hidden_dim_size(self) -> int:
         """Calculate the LLM hidden dimension size."""
         example = "test"
         inputs = self.tokenizer(example, padding=True, return_tensors="pt").to(
@@ -131,7 +131,7 @@ class CCMIL(VarAttention):
         """
         # [0] to convert list to string,
         # because all tiles in a bag have the same clinical context.
-        x, y, cc = batch["data"], batch["target"], batch["cc"][0]
+        x, y, cc = batch["data"], batch["target"], batch["cc"]
         y_hat, A = self(x, cc)
         loss = self.loss_fn(y_hat, y)
         return loss, y_hat, y, A