+++ dpat/utils/__init__.py
@@ -1,6 +1,12 @@
 """Provide utilities."""
 from .cudnn_auto_tuner import enable_cudnn_auto_tuner
+from .evaluation import compute_mean_and_confidence_interval
 from .float32_matmul_precision import set_float32_matmul_precision
 from .seed_all import seed_all
 
-__all__ = ["enable_cudnn_auto_tuner", "set_float32_matmul_precision", "seed_all"]
+__all__ = [
+    "enable_cudnn_auto_tuner",
+    "set_float32_matmul_precision",
+    "seed_all",
+    "compute_mean_and_confidence_interval",
+]